requirements.txt
Django==1.4.1
Pillow==1.7.7
South==0.7.6
anyjson==0.3.3
django-appconf==0.5
django-grappelli==2.4.2
django-imagekit==2.0.2
django-json-field==0.4.1
ipython==0.13
python-dateutil==2.1
six==1.2.0
